polish order example
