+++ example/src/main/java/org/apache/rocketmq/example/ordermessage/OrderedConsumer.java
@@ -17,7 +17,6 @@
 package org.apache.rocketmq.example.ordermessage;
 
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
 import org.apache.rocketmq.client.consumer.DefaultMQPushConsumer;
 import org.apache.rocketmq.client.consumer.listener.ConsumeOrderlyContext;
 import org.apache.rocketmq.client.consumer.listener.ConsumeOrderlyStatus;
@@ -26,35 +25,64 @@ import org.apache.rocketmq.client.exception.MQClientException;
 import org.apache.rocketmq.common.consumer.ConsumeFromWhere;
 import org.apache.rocketmq.common.message.MessageExt;
 
-public class Consumer {
+public class OrderedConsumer {
 
     public static void main(String[] args) throws MQClientException {
-        DefaultMQPushConsumer consumer = new DefaultMQPushConsumer("please_rename_unique_group_name_3");
+
+        DefaultMQPushConsumer consumer = new DefaultMQPushConsumer("please_rename_unique_group_name");
 
         consumer.setConsumeFromWhere(ConsumeFromWhere.CONSUME_FROM_FIRST_OFFSET);
 
-        consumer.subscribe("TopicTest", "TagA || TagC || TagD");
+        consumer.subscribe("OrderedMsgTest", "TagA || TagC || TagD");
 
         consumer.registerMessageListener(new MessageListenerOrderly() {
-            AtomicLong consumeTimes = new AtomicLong(0);
 
             @Override
             public ConsumeOrderlyStatus consumeMessage(List<MessageExt> msgs, ConsumeOrderlyContext context) {
-                context.setAutoCommit(false);
-                System.out.printf("%s Receive New Messages: %s %n", Thread.currentThread().getName(), msgs);
-                this.consumeTimes.incrementAndGet();
-                if ((this.consumeTimes.get() % 2) == 0) {
-                    return ConsumeOrderlyStatus.SUCCESS;
-                } else if ((this.consumeTimes.get() % 3) == 0) {
-                    return ConsumeOrderlyStatus.ROLLBACK;
-                } else if ((this.consumeTimes.get() % 4) == 0) {
-                    return ConsumeOrderlyStatus.COMMIT;
-                } else if ((this.consumeTimes.get() % 5) == 0) {
-                    context.setSuspendCurrentQueueTimeMillis(3000);
-                    return ConsumeOrderlyStatus.SUSPEND_CURRENT_QUEUE_A_MOMENT;
-                }
 
-                return ConsumeOrderlyStatus.SUCCESS;
+                /*
+                 * The default value of AutoCommit is true and RocketMQ will commit offset automatically
+                 * but if you set AutoCommit to false, the offset can only be commited by returning ConsumeOrderlyStatus.COMMIT
+                 */
+                context.setAutoCommit(false);
+
+                /*
+                 * Messages with the same orderID will be consumed sequentially
+                 */
+                System.out.printf("%s Receive New Messages: %s %n", Thread.currentThread().getName(), msgs);
+
+                /*
+                 * Returning ConsumeOrderlyStatus.SUCCESS means the messages was consumed successfully,
+                 * but the offset will not be committed if AutoCommit is false
+                 *
+                 * {@code
+                 * return ConsumeOrderlyStatus.SUCCESS;
+                 * }
+                 */
+
+                /*
+                 * Returning ConsumeOrderlyStatus.ROLLBACK means the messages after the commit point will be re-consumed
+                 *
+                 * {@code
+                 * return ConsumeOrderlyStatus.ROLLBACK;
+                 * }
+                 */
+
+                /*
+                 * Returning ConsumeOrderlyStatus.SUSPEND_CURRENT_QUEUE_A_MOMENT means
+                 * the message consumption of the current queue will be suspended for a while.
+                 *
+                 * {@code
+                 * context.setSuspendCurrentQueueTimeMillis(3000);
+                 * return ConsumeOrderlyStatus.SUSPEND_CURRENT_QUEUE_A_MOMENT;
+                 * }
+                 */
+
+
+                /*
+                 * Returning ConsumeOrderlyStatus.COMMIT means the max offset of the list will be committed
+                 */
+                return ConsumeOrderlyStatus.COMMIT;
             }
         });
